+++ starlette/routing.py
@@ -38,21 +38,6 @@ class Match(Enum):
     FULL = 2
 
 
-def iscoroutinefunction_or_partial(obj: typing.Any) -> bool:  # pragma: no cover
-    """
-    Correctly determines if an object is a coroutine function,
-    including those wrapped in functools.partial objects.
-    """
-    warnings.warn(
-        "iscoroutinefunction_or_partial is deprecated, "
-        "and will be removed in a future release.",
-        DeprecationWarning,
-    )
-    while isinstance(obj, functools.partial):
-        obj = obj.func
-    return inspect.iscoroutinefunction(obj)
-
-
 def request_response(func: typing.Callable) -> ASGIApp:
     """
     Takes a function or coroutine `func(request) -> response`,
@@ -742,36 +727,24 @@ class Router:
     def mount(
         self, path: str, app: ASGIApp, name: typing.Optional[str] = None
     ) -> None:  # pragma: nocover
-        """
-        We no longer document this API, and its usage is discouraged.
-        Instead you should use the following approach:
-
-        routes = [
-            Mount(path, ...),
-            ...
-        ]
-
-        app = Starlette(routes=routes)
-        """
-
+        warnings.warn(
+            "The 'mount' method is now deprecated, and will be removed in version "
+            "2.0.0. Refer to https://www.starlette.io/routing/#submounting-routes "
+            "for recommended approach.",
+            DeprecationWarning,
+        )
         route = Mount(path, app=app, name=name)
         self.routes.append(route)
 
     def host(
         self, host: str, app: ASGIApp, name: typing.Optional[str] = None
     ) -> None:  # pragma: no cover
-        """
-        We no longer document this API, and its usage is discouraged.
-        Instead you should use the following approach:
-
-        routes = [
-            Host(path, ...),
-            ...
-        ]
-
-        app = Starlette(routes=routes)
-        """
-
+        warnings.warn(
+            "The 'host' method is deprecated, and will be removed in version 2.0.0."
+            "Refer to https://www.starlette.io/routing/#host-based-routing for the "
+            "recommended approach.",
+            DeprecationWarning,
+        )
         route = Host(host, app=app, name=name)
         self.routes.append(route)
 
@@ -805,17 +778,12 @@ class Router:
         name: typing.Optional[str] = None,
         include_in_schema: bool = True,
     ) -> typing.Callable:  # pragma: nocover
-        """
-        We no longer document this decorator style API, and its usage is discouraged.
-        Instead you should use the following approach:
-
-        routes = [
-            Route(path, endpoint=..., ...),
-            ...
-        ]
-
-        app = Starlette(routes=routes)
-        """
+        warnings.warn(
+            "The `route` decorator is deprecated, and will be removed in version 2.0.0."
+            "Refer to https://www.starlette.io/routing/#http-routing for the "
+            "recommended approach.",
+            DeprecationWarning,
+        )
 
         def decorator(func: typing.Callable) -> typing.Callable:
             self.add_route(
@@ -832,17 +800,13 @@ class Router:
     def websocket_route(
         self, path: str, name: typing.Optional[str] = None
     ) -> typing.Callable:  # pragma: nocover
-        """
-        We no longer document this decorator style API, and its usage is discouraged.
-        Instead you should use the following approach:
-
-        routes = [
-            WebSocketRoute(path, endpoint=..., ...),
-            ...
-        ]
-
-        app = Starlette(routes=routes)
-        """
+        warnings.warn(
+            "The `websocket_route` decorator is deprecated, and will be removed in "
+            "version 2.0.0. Refer to "
+            "https://www.starlette.io/routing/#websocket-routing for the recommended "
+            "approach.",
+            DeprecationWarning,
+        )
 
         def decorator(func: typing.Callable) -> typing.Callable:
             self.add_websocket_route(path, func, name=name)
@@ -861,6 +825,13 @@ class Router:
             self.on_shutdown.append(func)
 
     def on_event(self, event_type: str) -> typing.Callable:  # pragma: nocover
+        warnings.warn(
+            "The `on_event` decorator is deprecated, and will be removed in version "
+            "2.0.0. Refer to https://www.starlette.io/events/#registering-events for "
+            "recommended approach.",
+            DeprecationWarning,
+        )
+
         def decorator(func: typing.Callable) -> typing.Callable:
             self.add_event_handler(event_type, func)
             return func