+++ tests/test_exceptions.py
@@ -1,5 +1,3 @@
-import warnings
-
 import pytest
 
 from starlette.exceptions import HTTPException, WebSocketException
@@ -147,16 +145,3 @@ def test_websocket_repr():
         repr(CustomWebSocketException(1013, reason="Something custom"))
         == "CustomWebSocketException(code=1013, reason='Something custom')"
     )
-
-
-def test_exception_middleware_deprecation() -> None:
-    # this test should be removed once the deprecation shim is removed
-    with pytest.warns(DeprecationWarning):
-        from starlette.exceptions import ExceptionMiddleware  # noqa: F401
-
-    with warnings.catch_warnings():
-        warnings.simplefilter("error")
-        import starlette.exceptions
-
-    with pytest.warns(DeprecationWarning):
-        starlette.exceptions.ExceptionMiddleware